\name{Constants}
\alias{MAX_CLUSTERS}
\alias{CLUSTER_NAMES}
\alias{CRITERION_MIN_OR_MAX}
\docType{data}
\title{
  ~ Constants define in the package ~
}
\description{
  Constants define in the package ~
}
\usage{
MAX_CLUSTERS
CLUSTERS_NAMES
CRITERION_MIN_OR_MAX
}

\value{
  MAX_CLUSTERS = 26
  CLUSTER_NAMES = paste("c",2:MAX_CLUSTERS,sep="")
  CRITERION_MIN_OR_MAX = c(calinski=1,test=1,test2=-1)

}
\source{
}
\references{
}
\examples{
#cat(MAX_CLUSTERS)
}
\keyword{datasets}
