\name{restaureRealData}
\alias{restaureRealData}
\alias{restaureRealData,LongData}
\alias{restaureRealData,LongData-method}

\title{
~ Function: restaureRealData ~
}
\description{
  This function revert the effect of \code{\link{scale}} by restauring
  the initial values of trajectories.
}
\usage{
restaureRealData(object)
}
\arguments{
  \item{object}{\code{[LongData]}: Object containnig trajectories to restaure.}
}

\details{
  This function revert the effect of \code{\link{scale}} by restauring
  the initial values of trajectories.
}
\value{
None. (this function change internaly the field of an object, it does not return any values.)
}
\references{
}
\author{
}
\note{
}


\seealso{
}
\examples{
}
